Adding testing and extracting detector stuff to its own file to someday make this a self contained module.
